refactor processupload
