+++ src/components/data-management/ProjectsListContainer.jsx
@@ -14,14 +14,15 @@ import FileUploadModal from './FileUploadModal';
 import { setActiveProject } from '../../redux/actions/projects';
 import PrettyTime from '../PrettyTime';
 
-import { createSample, updateSampleFile } from '../../redux/actions/samples';
+// import { createSample, updateSampleFile } from '../../redux/actions/samples';
+import processUpload from '../../utils/processUpload';
 
 const ProjectsListContainer = (props) => {
   const { height } = props;
   const dispatch = useDispatch();
 
-  const projects = useSelector((state) => state.projects);
   const samples = useSelector((state) => state.samples);
+  const projects = useSelector((state) => state.projects);
   const { activeProject } = projects.meta;
   const [deleteModalVisible, setDeleteModalVisible] = useState(false);
   const [uploadModalVisible, setUploadModalVisible] = useState(true);
@@ -37,41 +38,7 @@ const ProjectsListContainer = (props) => {
   };
 
   const uploadFiles = (filesList, sampleType) => {
-    const samplesMap = filesList.reduce((acc, file) => {
-      const sampleName = file.name.trim().replace(/[\s]{2,}/ig, ' ').split('/')[0];
-      const sampleUuid = Object.values(samples).filter(
-        (s) => s.name === sampleName
-          && s.projectUuid === projects.meta.activeProject,
-      )[0]?.uuid;
-
-      return {
-        ...acc,
-        [sampleName]: {
-          ...acc[sampleName],
-          uuid: sampleUuid,
-          files: {
-            ...acc[sampleName]?.files,
-            [sampleName]: file,
-          },
-        },
-      };
-    }, {});
-
-    Object.entries(samplesMap).forEach(async ([name, sample]) => {
-      // Create sample if not exists
-      if (!sample.uuid) {
-        // eslint-disable-next-line no-param-reassign
-        sample.uuid = await dispatch(createSample(projects.meta.activeProject, name, sampleType));
-      }
-
-      Object.values(sample.files).forEach((file) => {
-        dispatch(updateSampleFile(sample.uuid, {
-          ...file,
-          path: `${projects.meta.activeProject}/${file.name.replace(name, sample.uuid)}`,
-        }));
-      });
-    });
-
+    processUpload(filesList, sampleType, samples, activeProject, dispatch);
     setUploadModalVisible(false);
   };
